src/backend/python/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
flask-sqlalchemy>=3.1.1
flask-jwt-extended>=4.6.0
flask-bcrypt>=1.0.1
flask-limiter>=3.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pyserial>=3.5
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.2.0
joblib>=1.3.2
eventlet>=0.35.2
gunicorn>=21.2.0
